# openapi.yaml
openapi: 3.0.3
info:
  contact:
    email: info@librematch.org
    name: Libre:Match Team
    url: https://librematch.org
  description:
    "stable, reliable and feature-rich API for match information, leaderboards,\
    \ and tournament organisation for the Age of Empires multiplayer community"
  license:
    name: GNU AGPLv3
    url: https://www.gnu.org/licenses/agpl.txt
    x-last-modified: 1663947772699
  title: Libre:Match API
  version: 1.0.0-DRAFT
  x-logo:
    url: ""
servers:
  - description: Libre:Match Server v1
    url: https://api.librematch.org/v1
    x-last-modified: 1663947810574
security:
  - BasicAuth: []
  - X-Api-Key: []
paths:
...
  /users:
    delete:
      description: ""
      operationId: remove_all_users
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
          description: Unauthorized access
        "200":
          description: Operation succeeded
          x-last-modified: 1665056670316
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
          description: Unauthorized access
          x-last-modified: 1665072543221
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForbiddenError"
          description: Forbidden
          x-last-modified: 1665056800086
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
          description: Entity not found
          x-last-modified: 1665056674861
        "406":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotAcceptableError"
          description: Not acceptable
          x-last-modified: 1665056674861
        "429":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TooManyRequestsError"
          description: Too Many Requests
          x-last-modified: 1665056827121
      summary: Remove all users from our platform
      tags:
        - User Administration
    get:
      description: Authenticated Endpoint for Administrators or Libre:Match
      operationId: get_user_collection
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
          description: Unauthorized access
        "200":
          description: Operation succeeded
          x-last-modified: 1665056670316
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
          description: Unauthorized access
          x-last-modified: 1665072543221
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForbiddenError"
          description: Forbidden
          x-last-modified: 1665056800086
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
          description: Entity not found
          x-last-modified: 1665056674861
        "406":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotAcceptableError"
          description: Not acceptable
          x-last-modified: 1665056674861
        "429":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TooManyRequestsError"
          description: Too Many Requests
          x-last-modified: 1665056827121
      summary: Retrieve all users of our platform
      tags:
        - User Administration
    patch:
      description: ""
      operationId: partial_update_all_users
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
          description: Unauthorized access
        "200":
          description: Operation succeeded
          x-last-modified: 1665056670316
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
          description: Unauthorized access
          x-last-modified: 1665072543221
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForbiddenError"
          description: Forbidden
          x-last-modified: 1665056800086
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
          description: Entity not found
          x-last-modified: 1665056674861
        "406":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotAcceptableError"
          description: Not acceptable
          x-last-modified: 1665056674861
        "415":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsupportedMediaTypeError"
          description: Unsupported Media Type
          x-last-modified: 1665057248987
        "429":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TooManyRequestsError"
          description: Too Many Requests
          x-last-modified: 1665056827121
      summary: Partially update details for all user profiles
      tags:
        - User Administration
    post:
      description: ""
      operationId: create_new_user
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
          description: Unauthorized access
        "200":
          description: Operation succeeded
          x-last-modified: 1665056670316
        "201":
          description: Created
          x-last-modified: 1665072508175
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
          description: Unauthorized access
          x-last-modified: 1665072543221
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForbiddenError"
          description: Forbidden
          x-last-modified: 1665056800086
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
          description: Entity not found
          x-last-modified: 1665056674861
        "406":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotAcceptableError"
          description: Not acceptable
          x-last-modified: 1665056674861
        "415":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsupportedMediaTypeError"
          description: Unsupported Media Type
          x-last-modified: 1665057248987
        "429":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TooManyRequestsError"
          description: Too Many Requests
          x-last-modified: 1665056827121
      summary: Create a new user for our platform
      tags:
        - Authentication
        - Account Registration
        - User Administration
    put:
      description: ""
      operationId: bulk_update_all_users
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
          description: Unauthorized access
        "200":
          description: Operation succeeded
          x-last-modified: 1665056670316
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
          description: Unauthorized access
          x-last-modified: 1665072543221
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForbiddenError"
          description: Forbidden
          x-last-modified: 1665056800086
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
          description: Entity not found
          x-last-modified: 1665056674861
        "406":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotAcceptableError"
          description: Not acceptable
          x-last-modified: 1665056674861
        "415":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsupportedMediaTypeError"
          description: Unsupported Media Type
          x-last-modified: 1665057248987
        "429":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TooManyRequestsError"
          description: Too Many Requests
          x-last-modified: 1665056827121
      summary: Bulk update of all users of our platform
      tags:
        - User Administration
    summary: Endpoint for users of our API
    x-last-modified: 1663948638250
...
  /users/login:
    post:
      operationId: user_login
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
          description: Unauthorized access
        "200":
          description: Operation succeeded
          x-last-modified: 1665056670316
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
          description: Unauthorized access
          x-last-modified: 1665072543221
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForbiddenError"
          description: Forbidden
          x-last-modified: 1665056800086
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
          description: Entity not found
          x-last-modified: 1665056674861
        "406":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotAcceptableError"
          description: Not acceptable
          x-last-modified: 1665056674861
        "415":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsupportedMediaTypeError"
          description: Unsupported Media Type
          x-last-modified: 1665057248987
        "429":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TooManyRequestsError"
          description: Too Many Requests
          x-last-modified: 1665056827121
      tags:
        - Authentication
    summary: Login for users of our API
    x-last-modified: 1663948709646
  /users/logout:
    get:
      operationId: user_logout
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
          description: Unauthorized access
        "200":
          description: Operation succeeded
          x-last-modified: 1665056670316
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
          description: Unauthorized access
          x-last-modified: 1665072543221
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForbiddenError"
          description: Forbidden
          x-last-modified: 1665056800086
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
          description: Entity not found
          x-last-modified: 1665056674861
        "406":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotAcceptableError"
          description: Not acceptable
          x-last-modified: 1665056674861
        "429":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TooManyRequestsError"
          description: Too Many Requests
          x-last-modified: 1665056827121
      tags:
        - Authentication
    summary: Logout for users of our API
    x-last-modified: 1663948774846
...
  /users/validate:
    description:
      "Validate API tokens, can give information about how long a token\
      \ is still valid and help automating updating a token"
    get:
      operationId: get_user_validation_details
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
          description: Unauthorized access
        "200":
          description: Operation succeeded
          x-last-modified: 1665056670316
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
          description: Unauthorized access
          x-last-modified: 1665072543221
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForbiddenError"
          description: Forbidden
          x-last-modified: 1665056800086
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
          description: Entity not found
          x-last-modified: 1665056674861
        "406":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotAcceptableError"
          description: Not acceptable
          x-last-modified: 1665056674861
        "429":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TooManyRequestsError"
          description: Too Many Requests
          x-last-modified: 1665056827121
      tags:
        - Authentication
        - API Token Validation
    post:
      operationId: create_new_user_validation
      responses:
        default:
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
          description: Unauthorized access
        "200":
          description: Operation succeeded
          x-last-modified: 1665056670316
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnauthorizedError"
          description: Unauthorized access
          x-last-modified: 1665072543221
        "403":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ForbiddenError"
          description: Forbidden
          x-last-modified: 1665056800086
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundError"
          description: Entity not found
          x-last-modified: 1665056674861
        "406":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotAcceptableError"
          description: Not acceptable
          x-last-modified: 1665056674861
        "415":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnsupportedMediaTypeError"
          description: Unsupported Media Type
          x-last-modified: 1665057248987
        "429":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TooManyRequestsError"
          description: Too Many Requests
          x-last-modified: 1665056827121
      tags:
        - Authentication
        - API Token Validation
    summary: Validation for a user account and api tokens used for our API
    x-last-modified: 1663948876606
...
components:
...
  schemas:
...
    UnsupportedMediaTypeError:
      additionalProperties: false
      properties:
        message:
          title: message
          type: string
        code:
          maximum: 415
          minimum: 415
          title: code
          type: integer
      required:
        - code
        - message
      title: UnsupportedMediaTypeError
      type: object
      x-last-modified: 1665056674054
...
    TooManyRequestsError:
      additionalProperties: false
      properties:
        message:
          title: message
          type: string
        code:
          maximum: 429
          minimum: 429
          title: code
          type: integer
      required:
        - code
        - message
      title: TooManyRequestsError
      type: object
      x-last-modified: 1665056674054
...
    NotAcceptableError:
      additionalProperties: false
      properties:
        message:
          title: message
          type: string
        code:
          maximum: 406
          minimum: 406
          title: code
          type: integer
      required:
        - code
        - message
      title: NotAcceptableError
      type: object
      x-last-modified: 1665056674054
    NotFoundError:
      additionalProperties: false
      properties:
        message:
          title: message
          type: string
        code:
          maximum: 404
          minimum: 404
          title: code
          type: integer
      required:
        - code
        - message
      title: NotFoundError
      type: object
      x-last-modified: 1665056674054
    ForbiddenError:
      additionalProperties: false
      properties:
        message:
          title: message
          type: string
        code:
          maximum: 403
          minimum: 403
          title: code
          type: integer
      required:
        - code
        - message
      title: ForbiddenError
      type: object
      x-last-modified: 1665056674054
...
    UnauthorizedError:
      additionalProperties: false
      properties:
        message:
          maxLength: 32
          pattern: "^example-[0-9a-z]+$"
          title: Unauthorized access
          type: string
        code:
          maximum: 401
          minimum: 401
          title: code
          type: integer
      required:
        - code
        - message
      title: UnauthorizedError
      type: object
      x-last-modified: 1665056676293
...
  securitySchemes:
    BasicAuth:
      description: "Example: \n> Authorization: <Sample>"
      in: header
      name: BasicAuth
      type: apiKey
      x-last-modified: 1663947718456
    X-Api-Key:
      description: "Example: \n> X-Api-Key: ZGVtbzpwQDU1dzByZA=="
      in: header
      name: X-Api-Key
      type: apiKey
      x-last-modified: 1663947718456
    JWTAuth:
      bearerFormat: JWT
      description: "Example: \n> Authorization: Bearer <token>"
      scheme: bearer
      type: http
      x-last-modified: 1665071628782